requirements.txt
docx2txt==0.8
fastapi==0.108.0
langchain==0.1.0
langchain_google_genai==0.0.5
pdfplumber==0.10.3
protobuf==4.25.1
pydantic==2.5.3
python-dotenv==1.0.0
Requests==2.31.0
simplejson==3.19.2
streamlit==1.29.0
streamlit_chatbox==1.1.11
uvicorn==0.25.0
google-generativeai
chromadb
pysqlite3-binary
langchain-community
